requirements.txt
# Stock Anomaly Detector Dependencies

# Data fetching
yfinance>=0.2.36

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Technical indicators
pandas-ta>=0.3.14b

# Forecasting
prophet>=1.1.5

# Database
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.4

# Telegram alerts
python-telegram-bot>=20.7

# Dashboard
streamlit>=1.29.0
plotly>=5.18.0

# Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
